# docs/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/postalOffices":{"post":{"tags":["postal-office-controller"],"operationId":"registerPostalOffices","requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PostalOffice"}}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"object","additionalProperties":{"type":"string"}},{"type":"string"}]}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PostalOffice"}}}}}}}},"/parcels":{"get":{"tags":["parcel-controller"],"operationId":"getParcels","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"object","additionalProperties":{"type":"string"}},{"type":"string"}]}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Parcel"}}}}}}},"post":{"tags":["parcel-controller"],"operationId":"registerParcels","requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RegisterParcelDto"}}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"object","additionalProperties":{"type":"string"}},{"type":"string"}]}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Parcel"}}}}}}}},"/parcels/{parcel_id}/receive":{"post":{"tags":["parcel-controller"],"operationId":"receive","parameters":[{"name":"parcel_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReceiveParcelDto"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"object","additionalProperties":{"type":"string"}},{"type":"string"}]}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Parcel"}}}}}}},"/parcels/{parcel_id}/depart":{"post":{"tags":["parcel-controller"],"operationId":"depart","parameters":[{"name":"parcel_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DepartureParcelDto"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"object","additionalProperties":{"type":"string"}},{"type":"string"}]}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Transfer"}}}}}}},"/parcels/{parcel_id}/arrive":{"post":{"tags":["parcel-controller"],"operationId":"arrive","parameters":[{"name":"parcel_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArrivalParcelDto"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"object","additionalProperties":{"type":"string"}},{"type":"string"}]}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Transfer"}}}}}}},"/parcels/{parcel_id}/history":{"get":{"tags":["parcel-controller"],"operationId":"getParcelHistory","parameters":[{"name":"parcel_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"object","additionalProperties":{"type":"string"}},{"type":"string"}]}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ParcelHistoryDto"}}}}}}}},"components":{"schemas":{"PostalOffice":{"required":["address","index","name"],"type":"object","properties":{"index":{"pattern":"^[0-9]{6}$","type":"string"},"name":{"type":"string"},"address":{"type":"string"}}},"RegisterParcelDto":{"required":["acceptance_index","parcel_type","receiver_address","receiver_index","receiver_name"],"type":"object","properties":{"parcel_type":{"type":"string","enum":["Письмо","Посылка","Бандероль","Открытка"]},"acceptance_index":{"pattern":"^[0-9]{6}$","type":"string"},"receiver_address":{"type":"string"},"receiver_name":{"type":"string"},"receiver_index":{"pattern":"^[0-9]{6}$","type":"string"}}},"Parcel":{"required":["acceptancePostalOffice","acceptance_date","parcel_type","receiverPostalOffice","receiver_address","receiver_name","status"],"type":"object","properties":{"acceptancePostalOffice":{"$ref":"#/components/schemas/PostalOffice"},"receiverPostalOffice":{"$ref":"#/components/schemas/PostalOffice"},"currentPostalOffice":{"$ref":"#/components/schemas/PostalOffice"},"parcel_id":{"type":"integer","format":"int64"},"receiver_address":{"type":"string"},"receiver_name":{"type":"string"},"parcel_type":{"type":"string","enum":["Письмо","Посылка","Бандероль","Открытка"]},"status":{"type":"string","enum":["Принято в отделении связи","В пути между пунктами","В почтовом офисе","Ожидает в месте вручения","Получено"]},"acceptance_date":{"type":"string","format":"date-time"},"receive_date":{"type":"string","format":"date-time"}}},"ReceiveParcelDto":{"required":["receiver_index"],"type":"object","properties":{"receiver_index":{"pattern":"^[0-9]{6}$","type":"string"}}},"DepartureParcelDto":{"required":["arrival_index","departure_index"],"type":"object","properties":{"departure_index":{"pattern":"^[0-9]{6}$","type":"string"},"arrival_index":{"pattern":"^[0-9]{6}$","type":"string"}}},"Transfer":{"required":["arrival_index","departure_date","departure_index","parcel"],"type":"object","properties":{"parcel":{"$ref":"#/components/schemas/Parcel"},"transfer_id":{"type":"integer","format":"int64"},"departure_date":{"type":"string","format":"date-time"},"arrival_date":{"type":"string","format":"date-time"},"departure_index":{"$ref":"#/components/schemas/PostalOffice"},"arrival_index":{"$ref":"#/components/schemas/PostalOffice"}}},"ArrivalParcelDto":{"required":["arrival_index"],"type":"object","properties":{"arrival_index":{"pattern":"^[0-9]{6}$","type":"string"}}},"ParcelHistoryDto":{"type":"object","properties":{"parcel":{"$ref":"#/components/schemas/Parcel"},"transfers":{"type":"array","items":{"$ref":"#/components/schemas/Transfer"}}}}}}}